backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.21.1
python-multipart==0.0.6
jinja2
numpy
requests>=2.31.0
pillow>=10.0.0